🧪 test(xml): omit lxml round-trip from coverage gate
The free-threaded and 3.15 CI jobs ship no lxml wheel, so the round-trip
test importorskipped itself there, leaving its body and the localnames
helper uncovered and dropping the gate to 99%. Move it into a dedicated
differential file listed in coverage run.omit, matching the existing
xpath/soupsieve/css differential suites, so it still validates wherever
lxml installs without making the gate environment-dependent.
