+++ pyproject.toml
@@ -302,6 +302,7 @@ ini_options.filterwarnings = [
 run.omit = [
   "tests/benchmarks/*",                               # run only under --codspeed (the CI benchmark job), never in the coverage run
   "tests/convert/test_css_to_xpath_differential.py",
+  "tests/dom/test_parse_xml_differential.py",
   "tests/query/match/test_soupsieve_differential.py",
   "tests/query/xpath/test_xpath_differential.py",
 ]
